+++ app/airq/celery.py
@@ -3,7 +3,7 @@ from celery import signals
 from celery.schedules import crontab
 from kombu.utils.url import safequote
 
-from airq import settings
+from airq import config
 
 
 BEAT_SCHEDULE = {
@@ -20,26 +20,26 @@ def get_celery_logger():
     return get_task_logger(__name__)
 
 
-celery = Celery(settings.app.import_name)
+celery = Celery(config.app.import_name)
 celery.conf.update(
     accept_content=["application/json"],
     beat_schedule=BEAT_SCHEDULE,
     broker_url="sqs://{}:{}@".format(
-        safequote(settings.AWS_ACCESS_KEY_ID), safequote(settings.AWS_SECRET_ACCESS_KEY)
+        safequote(config.AWS_ACCESS_KEY_ID), safequote(config.AWS_SECRET_ACCESS_KEY)
     ),
     broker_transport_options={"region": "us-west-1"},
     result_serializer="json",
-    task_default_queue=f"celery-{settings.FLASK_ENV}",
+    task_default_queue=f"celery-{config.FLASK_ENV}",
     task_serializer="json",
     worker_enable_remote_control=False,
     worker_hijack_root_logger=False,
 )
-celery.conf.update(settings.app.config)
+celery.conf.update(config.app.config)
 
 
 class ContextTask(celery.Task):  # type: ignore
     def __call__(self, *args, **kwargs):
-        with settings.app.app_context():
+        with config.app.app_context():
             return self.run(*args, **kwargs)
 
     def on_failure(self, exc, task_id, args, kwargs, einfo):
@@ -54,7 +54,7 @@ celery.Task = ContextTask
 
 @signals.setup_logging.connect
 def setup_celery_logging(**kwargs):
-    dictConfig(settings.LOGGING_CONFIG)
+    dictConfig(config.LOGGING_CONFIG)
 
 
 celery.autodiscover_tasks(["airq"])
